# order-service/internal/domain/order.go
package domain

import (
	"slices"
	"time"
)

type OrderStatus string

const (
	StatusPending   OrderStatus = "pending"
	StatusConfirmed OrderStatus = "confirmed"
	StatusShipped   OrderStatus = "shipped"
	StatusDelivered OrderStatus = "delivered"
	StatusCancelled OrderStatus = "cancelled"
)

var validTransitions = map[OrderStatus][]OrderStatus{
	StatusPending:   {StatusConfirmed, StatusCancelled},
	StatusConfirmed: {StatusShipped, StatusCancelled},
	StatusShipped:   {StatusDelivered},
}

func (s OrderStatus) CanTransitionTo(next OrderStatus) bool {
	return slices.Contains(validTransitions[s], next)
}

type Order struct {
	ID         string
	UserID     string
	Status     OrderStatus
	TotalPrice float64
	Items      []OrderItem
	CreatedAt  time.Time
	UpdatedAt  time.Time
}

type OrderItem struct {
	ID        string
	OrderID   string
	ProductID string
	Quantity  int32
	UnitPrice float64
}